+++ .github/workflows/slack-check.yml
@@ -41,7 +41,7 @@ jobs:
         run: |
           set -o pipefail && 
           cd slack &&
-          ./gradlew --no-daemon assemble &&
+          ./gradlew --no-daemon assemble --scan &&
           cd ..
       - name: Build container images
         run: |
@@ -88,7 +88,7 @@ jobs:
         run: |
           set -o pipefail && 
           cd slack &&
-          ./gradlew --no-daemon --continue jsTest &&
+          ./gradlew --no-daemon --continue jsTest --scan &&
           cd ..
       - name: Generate test report
         uses: mikepenz/action-junit-report@v4
