+++ docker/run_local.sh
@@ -2,7 +2,7 @@
 set -e
 
 # Builds and runs the project locally with Docker
-echo "🛠 Build project" && ./gradlew assemble
+echo "🛠 Build project" && ./gradlew assemble --scan
 echo "🧹 Clean up old Docker resources" && (docker image prune -af)
 echo "🏁 Start local tunnel" && (ssh -tt -R codinglove.serveo.net:80:localhost:8080 serveo.net &)
 sleep 1
