+++ gradle-plugins/src/main/kotlin/com/gchristov/thecodinglove/gradleplugins/BasePlugin.kt
@@ -57,16 +57,3 @@ class BaseBrowserPlugin : BaseMultiplatformPlugin() {
         }
     }
 }
-
-fun Project.binaryRootDirectory(): Directory = layout.buildDirectory.dir("dist/js").get()
-
-fun Project.envSecret(key: String): String {
-    val propFile = file("./secrets.properties")
-    val properties = Properties()
-    properties.load(FileInputStream(propFile))
-    val property = properties.getProperty(key)
-    if (property.isNullOrBlank()) {
-        throw IllegalStateException("Required property is missing: property=$key")
-    }
-    return property
-}
